explicitly casted return

+++ ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: shorwood <shorwood@student.le-101.fr>      +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/07/09 22:08:58 by shorwood     #+#   ##    ##    #+#       */
-/*   Updated: 2018/10/09 16:20:21 by shorwood    ###    #+. /#+    ###.fr     */
+/*   Updated: 2018/10/15 21:06:43 by shorwood    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -16,7 +16,7 @@ int		ft_strcmp(const char *str, const char *cmp)
 	while (*str || *cmp)
 	{
 		if (*str != *cmp)
-			return ((unsigned char)*str - (unsigned char)*cmp);
+			return (int)((unsigned char)*str - (unsigned char)*cmp);
 		str++;
 		cmp++;
 	}
